Remove Combiner from LazySegmentTree implementation.
Unlike the non-lazy segment tree (see: segment-tree.h) which only supports
point updates, the lazy segment tree does not actually work with an arbitrary
combiner, because we need special logic to determine how to update an interior
node without updating the entire subtree when doing a range update.

The following logic only works for sums of integers:

  data[idx] += value * (std::min(j, end) - std::max(i, start));

But to compute e.g. a maximum, we need to apply the update to the child subtrees
and recompute the maximum. See codelib3/LazyMaxSegmentTree.cc for an example.